Extract to a separate file
