+++ src/language-js/print/do-while-statement.js
@@ -1,12 +1,10 @@
 import { group, hardline } from "../../document/index.js";
-import {
-  printClause,
-  printDoWhileStatementCondition,
-} from "./miscellaneous.js";
+import { printDoWhileStatementBody } from "./clause.js";
+import { printDoWhileStatementCondition } from "./miscellaneous.js";
 
 function printDoWhileStatement(path, options, print) {
   return [
-    group(["do", printClause(path, print)]),
+    group(["do", printDoWhileStatementBody(path, print)]),
     path.node.body.type === "BlockStatement" ? " " : hardline,
     "while (",
     printDoWhileStatementCondition(path, options, print),
