+++ src/language-js/print/for-statement.js
@@ -6,11 +6,11 @@ import {
   softline,
 } from "../../document/index.js";
 import { printDanglingComments } from "../../main/comments/print.js";
-import { printClause } from "./miscellaneous.js";
+import { printForStatementBody } from "./clause.js";
 
 function printForStatement(path, options, print) {
   const { node } = path;
-  const body = printClause(path, print);
+  const body = printForStatementBody(path, print);
 
   // We want to keep dangling comments above the loop to stay consistent.
   // Any comment positioned between the for statement and the parentheses
